Create #Day 10:
Code Description:
Introduction of Tuple:
The code creates a tuple tup containing a mix of integers, a string, and a boolean. It prints the type and length of the tuple, accesses specific elements using indexing (both positive and negative), and checks for the presence of the value 342, printing a confirmation if it's found. Finally, it slices the tuple to create a new tuple tup2, containing elements from index 1 to 3 (2, 76, and 342). Overall, the code showcases basic tuple operations, including creation, element access, membership testing, and slicing.